web-scraper/requirements.txt
python-dotenv
selenium
supabase
beautifulsoup4
webdriver-manager
requests
sentence_transformers